JSON.c, Wrapper.c, and  CameraControl.c
Json.c is theoretically done. Simply need to compile and test to see if it works as intended.

Wrapper file has included function templates and an include statement for wrapper.h to interface with the json methods made in JSON.c.

CameraControl.c has code from the adeept manual that I received when I purchased their kit for the pi. It uses the wiringPi GPIO mode that works really well with the RP3B+.